Set up native function calls within the Lua scripting New class LuaRegistry creates the tables to be referenced in the scripts LuaBindingMacros will be used to parse the code and generate files for intellisense as a build step

+++ src/Engine/Gameobject.h
@@ -19,7 +19,7 @@ public:
 	void SetPosition(const sf::Vector2f& position);
 	void SetPosition(float x, float y);
 
-	uint64_t GetID() const;
+	uint32_t GetID() const;
 
 	bool IsActive() const { return m_isActive; }
 
@@ -28,7 +28,7 @@ protected:
 
 private:
 	bool m_isActive;
-	uint64_t m_id;
+	uint32_t m_id;
 };
 
 #endif
